api/requirements.txt
Django==3.0.8
Pillow==5.0.0
asgiref==3.2.10
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
flake8==3.8.3
markdown==3.2.2
psycopg2==2.8.4
pytz==2020.1
sqlparse==0.3.1
django-jet2==1.0.12